[Fix] kafka-to-nexus requires run-start control topic
Contrary to its documentation, kafka-to-nexus now refuses to start writing a job if the start-message does not define a control topic.
Thankfully, a control topic which matches the already-in-use command topic passes muster there, and does not cause a topic-switch.
The mccode_plumber.writer.writer_start signature includes a control_topic entry, likely this will always be command_topic for our use.

The default writer job pool topic has been renamed to WriterPool, and object attributes from job to pool to try and disambiguate the idea of a 'job pool' and a 'job control' topic.
Contorl of the kafka-to-nexus execution includes setting the verbosity flag, which  doesn't seem to do anything, other inputs are now required at the management-level.

Forwarder verbosity is configurable now as well.

EFU configuration is now done via a class that can serialize and deserialize itself, including to and from a command-line-suitable string.
This is to support instrument which require multiple EFUs, each of which needs its own calibration and configuration information.

Changes in the orchestration routines are aimed at debugging and fixing the workflow needed by kafka-to-nexus.
With starting/stopping the file writing process seemingly working well, a check now prevents running the simulation if a the start-message appears to fail.